requirements.txt
python-igraph
cairocffi
numpy
scipy
matplotlib
click
